Revert Emulator::Audio to use doubles instead of floats * fixes ODR violations, but is slightly slower (378fps->376fps)

+++ bsnes/emulator/platform.hpp
@@ -17,7 +17,7 @@ struct Platform {
   virtual auto open(uint id, string name, vfs::file::mode mode, bool required = false) -> shared_pointer<vfs::file> { return {}; }
   virtual auto load(uint id, string name, string type, vector<string> options = {}) -> Load { return {}; }
   virtual auto videoFrame(const uint16* data, uint pitch, uint width, uint height, uint scale) -> void {}
-  virtual auto audioFrame(const float* samples, uint channels) -> void {}
+  virtual auto audioFrame(const double* samples, uint channels) -> void {}
   virtual auto inputPoll(uint port, uint device, uint input) -> int16 { return 0; }
   virtual auto inputRumble(uint port, uint device, uint input, bool enable) -> void {}
   virtual auto dipSettings(Markup::Node node) -> uint { return 0; }
